IssueID #17459 - Create RPM for MapProxy:
A Dockerfile and fpm command to create an RPM for MapProxy. This
Dockerfile requires gdal_rpm_el7 to work.

The resulting mapproxy RPM has dependencies on libraries (e.g. python-shapely)
which are not in the base CentOS repositories, but are in Fedora's EPEL. On
the other hand the libkml and gdal libraries have trouble with the EPEL. One
possible way to install the resulting RPMs is to do:

  sudo yum install -y epel-release
  sudo yum install -y --disablerepo=epel libkml-1.3.0-1.el7.x86_64.rpm gdal-2.1.2-1.el7.x86_64.rpm
  sudo yum install -y mapproxy-1.9.0-1.el7.x86_64.rpm